remove component: 'wms'
