+++ src/components/MapAddLayer.vue
@@ -29,143 +29,311 @@
         <!-- MODAL BODY -->
         <div class = "modal-body">
 
-          <!-- LAYER PROJECTION -->
-          <div class = "form-group" v-disabled = "['kmz', 'zip'].includes(layer.type)">
-            <label for="projection-layer" v-t = "'mapcontrols.add_layer_control.select_projection'"></label>
-            <select class = "form-control" id = "projection-layer" v-model = "layer.crs">
-              <option v-for = "option in options" :value = "option">{{option}}</option>
+          <div class="form-group">
+            <label>Layer type</label>
+            <select id="add-layer-type" class = "form-control" v-model="layer_type">
+              <option disabled :value="undefined">Choose type</option>
+              <option value="wms">WMS</option>
+              <option value="file">File</option>
             </select>
           </div>
 
-          <!-- LAYER POSITION -->
-          <div class = "form-group">
-            <label for="position-layer" v-t = "'layer_position.message'"></label>
-            <select class = "form-control" id = "position-layer" v-model = "position">
-              <option :value = "'top'" v-t = "'layer_position.top'"></option>
-              <option :value = "'bottom'" v-t = "'layer_position.bottom'"></option>
-            </select>
-          </div>
+          <hr>
 
-          <!-- PERSISTENT LAYER  -->
-          <div class = "form-group">
-            <label for="persistent-layer" v-t = "'mapcontrols.add_layer_control.persistent_data'"></label>
-            <select class = "form-control" id = "persistent-layer" v-model = "persistent">
-              <option :value = "false" v-t = "'no'"></option>
-              <option :value = "true" v-t = "'yes'"></option>
-            </select>
-            <small v-t = "'mapcontrols.add_layer_control.persistent_help'"></small>
-          </div>
-
-          <!-- LAYER COLOR  -->
-          <p v-t = "'mapcontrols.add_layer_control.select_color'" style = "font-weight: 700;"></p>
-          <chrome-picker
-            v-model = "layer.color"
-            @input  = "onChangeColor"
-            style   ="width:100%; margin:auto"
-          />
-
+          <!-- LOADING INDICATOR -->
           <bar-loader :loading = "loading"/>
 
-          <form id = "addcustomlayer">
-            <input
-              ref     = "input_file"
-              type    = "file"
-              title   = " "
-              @change = "onChangeFile($event)"
-              :accept = "accepted_extension">
-            <h4 v-t="'mapcontrols.add_layer_control.drag_layer'"></h4>
-            <h4
-              v-if  = "layer.name"
-              class = "skin-color"
-              style = "font-weight: bold">{{ layer.name }}</h4>
-            <div>
-              <i
-                :class      = "g3wtemplate.getFontClass('cloud-upload')"
-                class       = "fa-5x"
-                aria-hidden = "true">
-              </i>
-            </div>
-            <p style = "font-weight: bold">[.gml, .geojson, .kml, .kmz ,.gpx, .csv, .zip(shapefile)]</p>
-          </form>
+          <template v-if="'wms' === layer_type">
 
-          <div v-if = "csv_extension" style = "padding: 15px; border: 1px solid grey; border-radius: 3px">
-            <bar-loader :loading = "csv.loading"/>
-            <div class = "select_field">
-              <label
-                v-t = "'mapcontrols.add_layer_control.select_csv_separator'"
-                for = "g3w-select-field-layer">
-              </label>
-              <select
-                id      = "g3w-select-separator"
-                class   = "form-control"
-                v-model = "csv.separator"
+            <div v-if="!wms_panel">
+
+              <!-- WMS URL -->
+              <div class = "form-group">
+                <label for = "add_custom_url_wms_input" title = "required">
+                  URL <i style = "font-family: Monospace;color: var(--skin-color);">*</i>
+                </label>
+                <a
+                  :href  = "`https://g3w-suite.readthedocs.io/en/v3.7.x/g3wsuite_client.html#wms`"
+                  target = "_blank"
+                  style  = "float: right;"
+                  title  = "Docs"
                 >
-                  <option
-                    v-for  = "separator in csv.separators"
-                    :key   = "separator"
-                    :value = "separator">{{ separator }}</option>
+                  <i :class = "$fa('external-link')"></i>
+                </a>
+                <input
+                  id      = "add_custom_url_wms_input"
+                  v-model = "url"
+                  class   = "form-control"
+                  style   = "width: 100%; color:#000000;"
+                  placeholder="http://example.org/?&service=WMS&request=GetCapabilities"
+                />
+              </div>
+
+              <!-- WMS NAME -->
+              <div class = "form-group">
+                <label for = "add_custom_name_url_wms_input" title = "required">
+                  <span v-t = "'sidebar.wms.panel.label.name'"></span>
+                  <i style = "font-family: Monospace;color: var(--skin-color);">*</i>
+                </label>
+                <input
+                  id      = "add_custom_name_url_wms_input"
+                  v-model = "id"
+                  class   = "form-control"
+                  style   = "width: 100%; color:#000000;">
+              </div>
+
+              <!-- SUBMIT BUTTON -->
+              <button
+                v-disabled          = "!inputswmsurlvalid"
+                style               = "width: 100%;"
+                @click.prevent.stop = "addwmsurl"
+                class               = "btn btn-block skin-background-color"
+              ><b :class = "$fa('plus-square')"></b></button>
+
+              <!-- LIST OF WMS LAYERS (STORED ON SERVER) -->
+              <div
+                v-for = "({ id }) in adminwmsurls"
+                :key  = "wmsurl"
+                style = "
+                  display: flex;
+                  justify-content: space-between;
+                  align-items: center;
+                  border-bottom: 1px solid #ccc;
+                  padding-bottom: 3px;
+                "
+              >
+
+                <!-- WMS ID -->
+                <span style = "flex-grow: 1;">{{ id }}</span>
+
+                <!-- ADD NEW WMS LAYER -->
+                <b
+                  @click.stop = "showWmsLayersPanel(id)"
+                  style       = "color: var(--skin-color); padding: 5px; font-size: 1.3em;"
+                  :class      = "$fa('plus-square')"
+                ></b>
+
+              </div>
+
+              <!-- LIST OF WMS LAYERS (STORED ON LOCAL STORAGE) -->
+              <div
+                v-for = "({id, url}) in localwmsurls"
+                :key  = "id"
+                style = "border-bottom: 1px solid #ccc; padding-bottom: 3px;"
+              >
+                <div style = "display: flex; justify-content: space-between; align-items: center; padding-top: 3px">
+
+                  <!-- WMS NAME -->
+                  <b style = "flex-grow: 1;">{{ id }}</b>
+
+                  <!-- ADD NEW WMS LAYER -->
+                  <i
+                    @click.stop            = "showWmsLayersPanel(url)"
+                    v-t-tooltip:top.create = "'sidebar.wms.add_wms_layer'"
+                    :class                 = "'wms-icon-action ' + $fa('plus-square')"
+                    style                  = "color: var(--skin-color); padding: 3px; margin: 2px;"
+                  ></i>
+
+                  <!-- DELETE WMS -->
+                  <i
+                    @click.stop            = "deleteWmsUrl(id)"
+                    v-t-tooltip:top.create = "'sidebar.wms.delete_wms_url'"
+                    :class                 = "'wms-icon-action ' + $fa('trash')"
+                    style                  = "color: red; padding: 3px; margin: 2px;"
+                  ></i>
+
+                </div>
+
+                <!-- WMS URL -->
+                <small>{{ url }}</small>
+
+              </div>
+
+            </div>
+
+            <div v-else-if="wms_panel" v-disabled = "loading">
+
+              <h3 class = "skin-color g3w-wms-panel-title">{{title}}</h3>
+
+              <helpdiv v-if = "abstract" :message = "abstract" />
+
+              <!-- LAYERS NAME   -->
+              <label for = "g3w-wms-layers" v-t = "'sidebar.wms.panel.label.layers'"></label>
+              <select id = "g3w-wms-layers" :multiple = "true" :clear = "true" v-select2 = "'selectedlayers'">
+                <option v-for = "layer in layers" :key = "layer.name" :value = "layer.name">{{ layer.title }}</option>
+              </select>
+
+              <!-- EPSG PROJECTIONS   -->
+              <label for = "g3w-wms-projections" v-t = "'sidebar.wms.panel.label.projections'"></label>
+              <select id = "g3w-wms-projections" v-select2 = "'epsg'">
+                <option v-for = "proj in projections" :key = "proj" :value = "proj">{{ proj }}</option>
+              </select>
+
+              <!-- NAME OF LAYER TO SAVE -->
+              <label for = "g3w-wms-layer-name" v-t = "'sidebar.wms.panel.label.name'"></label>
+              <input id  = "g3w-wms-layer-name" class = "form-control" v-model = "name">
+
+              <div v-if  = "added" class = "g3w-wms-external-panel-layer-added-message"
+                v-t   = "'sidebar.wms.layer_id_already_added'">
+              </div>
+
+              <!-- LAYER POSITION -->
+              <div class = "form-group">
+                <label for = "position-layer" v-t = "'layer_position.message'"></label>
+                <select id = "position-layer" class = "form-control" v-model = "position">
+                  <option :value = "'top'"    v-t = "'layer_position.top'"></option>
+                  <option :value = "'bottom'" v-t = "'layer_position.bottom'"></option>
                 </select>
-                <div
-                  class="select_field"
-                  :class="{'g3w-disabled': !csv.headers || 0 === csv.headers.length }"
-                >
-                  <label
-                    for = "g3w-select-x-field"
-                    v-t = "'mapcontrols.add_layer_control.select_csv_x_field'">
-                  </label>
-                  <select
-                    class   = "form-control"
-                    id      = "g3w-select-x-field"
-                    v-model = "csv.x"
+              </div>
+
+            </div>
+
+          </template>
+
+          <template v-else-if="'file' === layer_type">
+
+            <!-- LAYER PROJECTION -->
+            <div class = "form-group" v-disabled = "['kmz', 'zip'].includes(layer.type)">
+              <label for="projection-layer" v-t = "'mapcontrols.add_layer_control.select_projection'"></label>
+              <select class = "form-control" id = "projection-layer" v-model = "layer.crs">
+                <option v-for = "option in options" :value = "option">{{option}}</option>
+              </select>
+            </div>
+
+            <!-- LAYER POSITION -->
+            <div class = "form-group">
+              <label for="position-layer" v-t = "'layer_position.message'"></label>
+              <select class = "form-control" id = "position-layer" v-model = "position">
+                <option :value = "'top'" v-t = "'layer_position.top'"></option>
+                <option :value = "'bottom'" v-t = "'layer_position.bottom'"></option>
+              </select>
+            </div>
+
+            <!-- PERSISTENT LAYER  -->
+            <div class = "form-group">
+              <label for="persistent-layer" v-t = "'mapcontrols.add_layer_control.persistent_data'"></label>
+              <select class = "form-control" id = "persistent-layer" v-model = "persistent">
+                <option :value = "false" v-t = "'no'"></option>
+                <option :value = "true" v-t = "'yes'"></option>
+              </select>
+              <small v-t = "'mapcontrols.add_layer_control.persistent_help'"></small>
+            </div>
+
+            <!-- LAYER COLOR  -->
+            <p v-t = "'mapcontrols.add_layer_control.select_color'" style = "font-weight: 700;"></p>
+            <chrome-picker
+              v-model = "layer.color"
+              @input  = "onChangeColor"
+              style   ="width:100%; margin:auto"
+            />
+
+            <form id = "addcustomlayer">
+              <input
+                ref     = "input_file"
+                type    = "file"
+                title   = " "
+                @change = "onChangeFile($event)"
+                :accept = "accepted_extension">
+              <h4 v-t="'mapcontrols.add_layer_control.drag_layer'"></h4>
+              <h4
+                v-if  = "layer.name"
+                class = "skin-color"
+                style = "font-weight: bold">{{ layer.name }}</h4>
+              <div>
+                <i
+                  :class      = "g3wtemplate.getFontClass('cloud-upload')"
+                  class       = "fa-5x"
+                  aria-hidden = "true">
+                </i>
+              </div>
+              <p style = "font-weight: bold">[.gml, .geojson, .kml, .kmz ,.gpx, .csv, .zip(shapefile)]</p>
+            </form>
+
+            <div v-if = "csv_extension" style = "padding: 15px; border: 1px solid grey; border-radius: 3px">
+              <bar-loader :loading = "csv.loading"/>
+              <div class = "select_field">
+                <label
+                  v-t = "'mapcontrols.add_layer_control.select_csv_separator'"
+                  for = "g3w-select-field-layer">
+                </label>
+                <select
+                  id      = "g3w-select-separator"
+                  class   = "form-control"
+                  v-model = "csv.separator"
                   >
+                    <option
+                      v-for  = "separator in csv.separators"
+                      :key   = "separator"
+                      :value = "separator">{{ separator }}</option>
+                  </select>
+                  <div
+                    class="select_field"
+                    :class="{'g3w-disabled': !csv.headers || 0 === csv.headers.length }"
+                  >
+                    <label
+                      for = "g3w-select-x-field"
+                      v-t = "'mapcontrols.add_layer_control.select_csv_x_field'">
+                    </label>
+                    <select
+                      class   = "form-control"
+                      id      = "g3w-select-x-field"
+                      v-model = "csv.x"
+                    >
+                      <option
+                        v-for  = "header in csv.headers"
+                        :key   = "header"
+                        :value = "header">{{ header }}</option>
+                    </select>
+                  </div>
+                  <div
+                    class  = "select_field"
+                    :class = "{ 'g3w-disabled': !csv.headers || 0 === csv.headers.length }"
+                  >
+                    <label
+                      v-t = "'mapcontrols.add_layer_control.select_csv_y_field'"
+                      for = "g3w-select-y-field">
+                    </label>
+                    <select
+                      class   = "form-control"
+                      id      = "g3w-select-y-field"
+                      v-model = "csv.y"
+                    >
                     <option
                       v-for  = "header in csv.headers"
                       :key   = "header"
                       :value = "header">{{ header }}</option>
-                  </select>
-                </div>
-                <div
-                  class  = "select_field"
-                  :class = "{ 'g3w-disabled': !csv.headers || 0 === csv.headers.length }"
-                >
-                  <label
-                    v-t = "'mapcontrols.add_layer_control.select_csv_y_field'"
-                    for = "g3w-select-y-field">
-                  </label>
-                  <select
-                    class   = "form-control"
-                    id      = "g3w-select-y-field"
-                    v-model = "csv.y"
-                  >
-                  <option
-                    v-for  = "header in csv.headers"
-                    :key   = "header"
-                    :value = "header">{{ header }}</option>
-                  </select>
-                </div>
+                    </select>
+                  </div>
+              </div>
             </div>
-          </div>
 
-          <div
-            class  = "select_field"
-            :class = "{ 'g3w-disabled': !fields || 0 === fields.length }"
-          >
-            <label
-              v-t = "'mapcontrols.add_layer_control.select_field_to_show'"
-              for = "g3w-select-field-layer">
-            </label>
-            <select
-              class   = "form-control"
-              id      = "g3w-select-field-layer"
-              v-model = "field"
+            <div
+              class  = "select_field"
+              :class = "{ 'g3w-disabled': !fields || 0 === fields.length }"
             >
-              <option :value = "null">---</option>
-              <option
-                v-for  = "field in fields"
-                :key   = "field"
-                :value = "field">{{ field }}</option>
-            </select>
-          </div>
+              <label
+                v-t = "'mapcontrols.add_layer_control.select_field_to_show'"
+                for = "g3w-select-field-layer">
+              </label>
+              <select
+                class   = "form-control"
+                id      = "g3w-select-field-layer"
+                v-model = "field"
+              >
+                <option :value = "null">---</option>
+                <option
+                  v-for  = "field in fields"
+                  :key   = "field"
+                  :value = "field">{{ field }}</option>
+              </select>
+            </div>
+
+          </template>
+
+        </div>
+
+        <!-- MODAL FOOTER -->
+        <div class = "modal-footer">
 
           <div
             v-if  = "error_message"
@@ -173,24 +341,37 @@
             v-t   = "error_message">
           </div>
 
-          <!-- MODAL FOOTER -->
-          <div class = "modal-footer">
-            <button
-              v-t         = "'add'"
-              type        = "button"
-              class       = "btn btn-success pull-left"
-              @click.stop = "addLayer"
-              :disabled   = "!add">
-            </button>
-            <button
-              v-t          = "'close'"
-              type         = "button"
-              class        = "btn btn-default"
-              data-dismiss = "modal">
-            </button>
-          </div>
+          <!-- ERROR NOTICE -->
+          <div v-if      = "status.error" class = "g3w-add-wms-url-message g3w-wmsurl-error">{{ $t('server_error') }}</div>
+          <div v-else-if = "status.added" class = "g3w-add-wms-url-message g3w-wmsurl-already-added">&#x26A0;&#xFE0F; {{ $t('sidebar.wms.url_already_added') }}</div>
 
-        </div>
+          <button
+            v-t          = "'close'"
+            type         = "button"
+            class        = "btn btn-default"
+            data-dismiss = "modal"
+          ></button>
+
+          <!-- SUBMIT BUTTON -->
+          <button
+            v-if        = "'wms' === layer_type && wms_panel"
+            v-t         = "'add'"
+            type        = "button"
+            class       = "btn btn-success"
+            @click.stop = "addWMSlayer({ url, position, epsg, layers: selectedlayers, name: name && name.trim() || undefined })"
+            v-disabled  = "0 === selectedlayers.length"
+          ></button>
+  
+          <button
+            v-if        = "'file' === layer_type"
+            v-t         = "'add'"
+            type        = "button"
+            class       = "btn btn-success"
+            @click.stop = "addLayer"
+            :disabled   = "!add"
+          ></button>
+
+          </div>
 
       </div>
     </div>
@@ -204,22 +385,11 @@ import { EPSG }                      from 'g3w-constants';
 import ApplicationState              from 'store/application';
 import Projections                   from 'store/projections';
 import GUI                           from 'services/gui';
+import DataRouterService             from 'services/data';
 import { createVectorLayerFromFile } from 'utils/createVectorLayerFromFile';
+import { getUniqueDomId }            from 'utils/getUniqueDomId';
 
 const SUPPORTED_FORMAT = ['zip','geojson', 'GEOJSON',  'kml', 'kmz', 'KMZ', 'KML', 'json', 'gpx', 'gml', 'csv'];
-const CSV_SEPARATORS   = [',', ';'];
-
-//Vue color componet
-ChromeComponent.mounted = async function() {
-    await this.$nextTick();    // remove all the tihing that aren't useful
-    $('.vue-color__chrome__toggle-btn').remove();
-    $('.vue-color__editable-input__label').remove();
-    $('.vue-color__chrome__saturation-wrap').css('padding-bottom','100px');
-    $('.vue-color__chrome').css({
-      'box-shadow': '0 0 0 0',
-      'border': '1px solid #97A1A8'
-    });
-};
 
 export default {
 
@@ -236,12 +406,19 @@ export default {
     }
 
     return {
+      layer_type:         undefined,
+      wms_panel:          false,
+      adminwmsurls:       this.$options.wmsurls || ApplicationState.project.wmsurls || [],
+      localwmsurls:       [], // array of object {id, url}
+      url:                null,
+      id:                 null,
+      status:             { error: false, added: false, },
       vectorLayer:        null,
       options:            EPSG,
       error_message:      '',
-      position:           'top',
+      position:           'top', // layer position on map
       persistent:         false,
-      loading:            false,
+      loading:            false, // loading reactive status
       fields:             [],
       field:              null,
       accepted_extension: SUPPORTED_FORMAT.map(f => `.${f}`).join(','),
@@ -251,8 +428,8 @@ export default {
         headers:     [],
         x:           null,
         y:           null,
-        separators : CSV_SEPARATORS,
-        separator:   CSV_SEPARATORS[0],
+        separators : [',', ';'],
+        separator:   ',',
       },
       layer: {
         name:   null,
@@ -274,12 +451,25 @@ export default {
         title:    null,
         id:       null,
         external: true,
-      }
+      },
+      name:           undefined,  // name of saved layer
+      title:          null,       // title of layer
+      abstract:       null,       // abstract
+      map_formats:    [],         // map formats
+      info_formats:   [],         // info formats
+      methods:        [],         // @since 3.9.0
+      layers:         [],         // Array of layers
+      selectedlayers: [],         // Selected layers
+      projections:    [],         // projections
+      epsg:           null,       // choose epsg project
+      added:          false,      // added layer (Boolean)
     }
   },
+
   components: {
-    'chrome-picker': ChromeComponent
+    'chrome-picker': ChromeComponent,
   },
+
   computed: {
 
     /**
@@ -294,7 +484,24 @@ export default {
      */
     add() {
       return this.layer.data || this.csv.valid;
-    }
+    },
+
+    /**
+     * @returns {false|*|boolean}
+     */
+    inputswmsurlvalid() {
+      return (
+        (
+          this.url !== null &&
+          this.url.trim() &&
+          this.url && this.url.match(/(http(s)?:\/\/.)?(www\.)?[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)/g) // whether is a valid url
+        ) &&
+        (
+          this.id !== null &&
+          this.id.trim()
+        )
+      )
+    },
 
   },
 
@@ -423,7 +630,7 @@ export default {
           persistent: !!this.persistent,
         });
         $(this.$refs.modal_addlayer).modal('hide');
-        this.clear();
+        this.clearFile();
       } catch(e) {
         this.error_message = 'sdk.errors.add_external_layer';
       }
@@ -434,7 +641,7 @@ export default {
     /**
      * @since 3.8.0
      */
-    clear() {
+    clearFile() {
       this.error_message = '';
       this.loading       = false;
       this.layer.name    = null;
@@ -450,6 +657,340 @@ export default {
       this.csv.valid     = false;
     },
 
+    clearPanel() {
+      this.wms_panel     = false;
+      this.selectedlayers = [];
+      this.name           = null;
+      this.loading        = false;
+    },
+
+    /**
+     * Add new WMS url
+     * 
+     * @param { Object } wms
+     * @param { string } wms.id
+     * @param { string } wms.url
+     * 
+     * @returns {*}
+     */
+      async addNewUrl(wms) {
+      const found  = this.localwmsurls.find(l => l.url == wms.url || l.id == wms.id);
+      const status = { error: false, added: !!found };
+      // when url is not yet added
+      if (!found) {
+        try {
+          const response = await this.getWMSLayers(wms.url);
+          // skip on invalid response
+          if (!response.result) {
+            throw 'invalid response';
+          }
+          const data = this.getLocalWMSData();
+          this.localwmsurls.push(wms);
+          data.urls = this.localwmsurls;
+          this.updateLocalWMSData(data);
+          response.wmsurl = wms.url;
+          this._showWmsLayersPanel(response);
+        } catch(e) {
+          console.warn(e);
+          status.error = true;
+        }
+      }
+      return status;
+    },
+
+    /**
+     * @returns { Promise<void> }
+     */
+    async addwmsurl() {
+      this.loading           = true;
+      const { error, added } = await this.addNewUrl({ url: this.url, id: this.id });
+      this.status.error      = error;
+      this.status.added      = added;
+      this.loading           = false;
+    },
+
+    /**
+     * Delete url from local storage
+     * 
+     * @param id
+     */
+    deleteWmsUrl(id) {
+      this.localwmsurls       = this.localwmsurls.filter(l => id !== l.id);
+      const data              = this.getLocalWMSData();
+      data.urls               = this.localwmsurls;
+      this.updateLocalWMSData(data);
+    },
+
+    /**
+     * Add external WMS layer to map
+     * 
+     * @param { Object } wms
+     * @param { string } wms.url
+     * @param { string } wms.name
+     * @param wms.epsg
+     * @param wms.position
+     * @param wms.opacity
+     * @param wms.visible
+     * @param wms.layers
+     *
+     * @returns {Promise<unknown>}
+     */
+    _addExternalWMSLayer({
+      url,
+      layers,
+      name,
+      epsg     = GUI.getService('map').getEpsg(),
+      position = 'top',
+      opacity,
+      visible  = true
+    } = {}) {
+      const map             = GUI.getService('map');
+      const { RasterLayer } = require('map/layers/imagelayer');
+      const projection      = ol.proj.get(epsg);
+
+      const promise = new Promise((res, rej) => {
+        const wmslayer = new RasterLayer({ id: name || getUniqueDomId(), layers, projection, url });
+        const olLayer  = wmslayer.getOLLayer();
+        olLayer.getSource().once('imageloadend', res);
+        olLayer.getSource().once('imageloaderror', rej);
+        map.addExternalLayer(wmslayer, { position, opacity, visible });
+      });
+    
+      return promise;
+    },
+
+    /**
+     * Check if a layer is already added to map
+     * 
+     * @param { Object } wms
+     * @param { string } wms.url
+     * @param { string } wms.name
+     * @param { string } wms.epsg
+     * @param { string} wms.position
+     * @param wms.opacity
+     * @param wms.layers
+     * @param {Boolean } wms.visible
+     * 
+     * @returns { Promise<void> }
+     */
+    async addWMSlayer({
+      url,
+      name    = `wms_${getUniqueDomId()}`,
+      layers  = [],
+      epsg,
+      position,
+      visible = true,
+      opacity = 1,
+    } = {}) {
+
+      try {
+        // check if WMS already added (by name)
+        const data = this.getLocalWMSData();
+
+        if (this.wms_panel) {
+          const wms = data.wms[this.url];
+          this.added = wms && wms.some(w => w.layers.length === this.selectedlayers.length
+            ? this.selectedlayers.every(l => w.layers.includes(l))
+            : undefined
+          );
+          if (this.added) {
+            console.warn('WMS Layer already added');
+            return;
+          }
+          this.loading = true;
+        }
+
+        const config = { url, name, layers, epsg, position, visible, opacity };
+
+        if (undefined === data.wms[url]) {
+          data.wms[url] = [config];
+        } else {
+          data.wms[url].push(config);
+        }
+
+        this.updateLocalWMSData(data);
+
+        try {
+          await this._addExternalWMSLayer(config);
+        } catch(e) {
+          console.warn(e);
+          GUI.getService('map').removeExternalLayer(name);
+          this.deleteWms(name);
+          setTimeout(() => { GUI.showUserMessage({ type: 'warning', message: 'sidebar.wms.layer_add_error' }) });
+        }
+      } catch(e) {
+        console.warn(e);
+      }
+      if (this.wms_panel) {
+        this.clearPanel();
+        $('#modal-addlayer').modal('hide');
+      }
+    },
+
+    /**
+     * Get data of wms url from server
+     * 
+     * @param { string } url
+     */
+    async getWMSLayers(url) {
+      try {
+        return await DataRouterService.getData('ows:wmsCapabilities', { inputs: { url }, outputs: false });
+      } catch(e) {
+        console.warn(e);
+      }
+      return {
+        result:       false,
+        layers:       [],
+        info_formats: [], // @deprecated since 3.9.0 (inside methods)
+        abstract:     null,
+        methods:      [], // @since 3.9.0
+        map_formats:  [], // @deprecated since 3.9.0 (inside methods)
+        title:        null
+      };
+    },
+
+    /**
+     * ORIGINAL SOURCE: src/app/gui/wms/vue/panel/wmslayerspanel.js@3.8.15
+     * 
+     * show add wms layers to wms panel
+     * 
+     * @param config
+     * 
+     * @returns { WmsLayersPanel }
+     */
+    _showWmsLayersPanel(config = {}) {
+      this.wms_panel = true;
+      this.wms_config = config;
+      const {
+        layers,
+        title,
+        abstract,
+        methods, // @since 3.9.0
+        wmsurl,
+      } = this.wms_config;
+
+      /** URL of wms */
+      try {
+        this.url = methods.GetMap.urls.find(u => 'Get' === u.type).url;
+      } catch(e) {
+        console.warn(e);
+        this.url = wmsurl;
+      }
+
+      /** Title of wms */
+      this.title = title;
+      
+      /** Abstract of wms */
+      this.abstract = abstract;
+
+      /** Store for each layer name projection info */
+      this.layerProjections = {};
+
+      /* try to check if projection */
+      layers
+        .forEach(({ name, crss, title }) => {
+          this.layerProjections[name] = {
+            title,
+            crss: crss.map(crs => { Projections.get(crs); return `EPSG:${crs.epsg}`; }).sort(),
+          };
+        });
+
+      /** Layers of wms */
+      this.layers = layers;
+    },
+
+    /**
+     * Load WMS Data from server and show WMS Layers Panel
+     * 
+     * @param url
+     * 
+     * @returns { Promise<void> }
+     */
+    async showWmsLayersPanel(url) {
+      let error = false;
+      let added = false;
+      try {
+        this.loading = true;
+        const d = await this.getWMSLayers(url);
+        error = !d.result;
+        if (!error) {
+          d.wmsurl = url;
+          this._showWmsLayersPanel(d);
+        }
+      } catch(e) {
+        console.warn(e);
+      } finally {
+        this.status.error = error;
+        this.status.added = added;
+        this.loading      = false;
+      }
+    },
+
+    /**
+     * Delete WMS by name
+     * 
+     * @param name
+     */
+    deleteWms(name) {
+      const data = this.getLocalWMSData();
+      Object.keys(data.wms).find(url => {
+        const i = data.wms[url].findIndex(w => w.name == name);
+        /** @TODO add description */
+        if (-1 !== i) {
+          data.wms[url].splice(i, 1);
+        }
+        /** @TODO add description */
+        if (-1 !== i && 0 == data.wms[url].length) {
+          delete data.wms[url];
+        }
+        return true;
+      });
+      this.updateLocalWMSData(data);
+    },
+
+    /**
+     * Change config of storage layer options as position, opacity
+     */
+    changeLayerData(name, attr = {}) {
+      const data = this.getLocalWMSData();
+      Object
+        .keys(data.wms)
+        .find(url => {
+          const i = data.wms[url].findIndex(l => l.name == name);
+          if (-1 !== i) {
+            data.wms[url][i][attr.key] = attr.value;
+            return true;
+          }
+        });
+      this.updateLocalWMSData(data);
+    },
+
+    /**
+     * Get local storage wms data based on current projectId
+     * 
+     * @returns {*}
+     */
+    getLocalWMSData() {
+      const item = window.localStorage.getItem('externalwms');
+      return ((item ? JSON.parse(item) : undefined) || {})[ApplicationState.project.getId()];
+    },
+
+    /**
+     * Update local storage data based on changes
+     * 
+     * @param data
+     */
+    updateLocalWMSData(data) {
+      const item = window.localStorage.getItem('externalwms');
+      const alldata = (item ? JSON.parse(item) : undefined) || {};
+      alldata[ApplicationState.project.getId()] = data;
+      try {
+        window.localStorage.setItem('externalwms', JSON.stringify(alldata));
+      } catch(e) {
+        console.warn(e);
+      }
+    },
+
   },
 
   watch: {
@@ -462,6 +1003,39 @@ export default {
       if (![undefined, null].includes(value)) { this.layer.data.y = value }
     },
 
+    /**
+     * Handle selected layers change  
+     */
+    selectedlayers(layers = []) {
+      if (0 === layers.length) {             // Reset epsg and projections to initial values
+        this.epsg        = null;
+        this.projections = [];
+      } else if (1 === layers.length) { // take first layer selected supported crss
+        this.epsg        = this.layerProjections[layers[0]].crss[0];
+        this.projections = this.layerProjections[layers[0]].crss;
+      } else {                          // get projections by name
+        const name = layers[layers.length -1];
+        this.projections = this.projections.filter(p => this.layerProjections[name].crss.includes(p));
+      }
+    },
+
+    /**
+     * @returns { Promise<void> }
+     */
+    async epsg() {
+      await this.$nextTick();
+      // Get layers that has current selected epsg projection
+      this.layers = (null === this.epsg)
+        ? this.wms_config.layers
+        : this.layers.filter(({ name }) => this.layerProjections[name].crss.includes(this.epsg))
+    },
+
+    layer_type(type) {
+      if ('file' == type) {
+        this.clearPanel();
+      }
+    }
+
   },
 
   created() {
@@ -470,48 +1044,125 @@ export default {
 
   async mounted() {
     await this.$nextTick();
+
     $('#modal-addlayer').modal('hide');
-    $('#modal-addlayer').on('hide.bs.modal',  () => this.clear() );
+    $('#modal-addlayer').on('hide.bs.modal',  () => {
+      this.layer_type = undefined;
+      this.clearFile();
+      this.clearPanel();
+    });
+
+    // Load WMS urls from local storage
+
+    await GUI.isReady();
+
+    const map = GUI.getService('map');
+
+    await map.isReady();
+
+    this.deleteWms = this.deleteWms.bind(this);
+
+    map.on('remove-external-layer', this.deleteWms);
+
+    let data = this.getLocalWMSData();
+
+    if (undefined === data) {
+      data = {
+        urls: [], // unique url for wms
+        wms:  {}, // object contains url as a key and array of layers bind to url
+      };
+      this.updateLocalWMSData(data);
+    }
+
+    setTimeout(() => {
+      map.on('change-layer-position-map', ({ id: name, position } = {}) => this.changeLayerData(name, { key: 'position', value: position }));
+      map.on('change-layer-opacity',      ({ id: name, opacity } = {})  => this.changeLayerData(name, { key: 'opacity',  value: opacity }));
+      map.on('change-layer-visibility',   ({ id: name, visible } = {})  => this.changeLayerData(name, { key: 'visible',  value: visible }));
+
+      // load eventually data
+      Object.keys(data.wms).forEach(url => { data.wms[url].forEach(d => this._addExternalWMSLayer({ url, ...d })); });
+    });
+
+    this.localwmsurls = data.urls;
   },
 
   beforeDestroy() {
-    this.clear();
+    this.clearFile();
+    this.clearPanel();
     $('#modal-addlayer').modal('hide')
     $('#modal-addlayer').remove();
+
+    GUI.getService('map').off('remove-external-layer', this.deleteWms);
+    this.$data = null;
   },
 
 };
 </script>
 
 <style scoped>
-#addcustomlayer {
-  margin: 10px 0 10px 0px;
-  position: relative;
-  border: 2px dashed #97A1A8;
-}
+  #addcustomlayer {
+    margin: 10px 0 10px 0px;
+    position: relative;
+    border: 2px dashed #97A1A8;
+  }
 
-#addcustomlayer p,
-#addcustomlayer h4 {
-  text-align: center;
-  line-height: 30px;
-  color: #97A1A8;
-  font-family: Arial;
-}
+  #addcustomlayer p,
+  #addcustomlayer h4 {
+    text-align: center;
+    line-height: 30px;
+    color: #97A1A8;
+    font-family: Arial;
+  }
 
-#addcustomlayer div {
-  text-align: center;
-  line-height: 30px;
-  color: #97A1A8;
-}
+  #addcustomlayer div {
+    text-align: center;
+    line-height: 30px;
+    color: #97A1A8;
+  }
 
-#addcustomlayer input{
-  position: absolute;
-  margin: 0;
-  padding: 0;
-  width: 100%;
-  height: 100%;
-  outline: none;
-  opacity: 0;
-  cursor: pointer;
-}
+  #addcustomlayer input{
+    position: absolute;
+    margin: 0;
+    padding: 0;
+    width: 100%;
+    height: 100%;
+    outline: none;
+    opacity: 0;
+    cursor: pointer;
+  }
+  .g3w-add-wms-url-message{
+    font-weight: bold;
+    color: #000000;
+  }
+  .g3w-wmsurl-error {
+    background-color: red;
+  }
+  .g3w-wmsurl-already-added {
+    color: inherit;
+    font-weight: normal;
+    display: inline-block;
+  }
+  #add_custom_url_wms_input::placeholder {
+    font-size: 85%;
+    opacity: .5;
+  }
+  .wms-icon-action {
+    font-weight: bold;
+    font-size: 1.3em;
+    cursor: pointer;
+  }
+  .g3w-wms-panel-title {
+    font-size: 1.2em;
+    font-weight: bold;
+    margin-bottom: 10px;
+  }
+  button.wms-add-layer-button {
+    width: 100%;
+    margin-top: 10px;
+  }
+  .g3w-wms-external-panel-layer-added-message {
+    font-weight: bold;
+    color: red;
+    margin: 5px 0;
+  }
 </style>
